BOCM-52: PR amends - refactoring to move data manipulation out of controllers and services into views
