+++ apps/web/src/server/views/_includes/macros/make-appeal-site-html.njk
@@ -1,0 +1,12 @@
+{% macro makeAppealSiteHtml(appealSiteData) %}
+	{%- set numRendered = 0 -%}
+	{%- for key, value in appealSiteData -%}
+		{%- if value -%}
+			{%- if numRendered > 0 -%}
+				<br />
+			{%- endif -%}
+			{{ value }}
+			{%- set numRendered = numRendered + 1 -%}
+		{%- endif -%}
+	{%- endfor -%}
+{% endmacro %}
